+++ app/Importers/OglImporter.php
@@ -2,8 +2,6 @@
 
 namespace App\Importers;
 
-use App\Repositories\IFileRepository;
-use Illuminate\Translation\Translator;
 use Illuminate\Support\Str;
 
 class OglImporter extends AbstractImporter
@@ -29,6 +27,8 @@ class OglImporter extends AbstractImporter
         'state_edition:sk' => 'Původnost',
         'title:cs' => 'Titul',
         'title:sk' => 'Titul',
+        'current_location:cs' => 'AktLokace',
+        'current_location:sk' => 'AktLokace',
     ];
 
     protected $defaults = [
@@ -38,15 +38,16 @@ class OglImporter extends AbstractImporter
 
     public static $name = 'ogl';
 
-    public function __construct(IFileRepository $repository, Translator $translator) {
-        parent::__construct($repository, $translator);
+    protected function init()
+    {
         $this->sanitizers[] = function ($value) {
             return empty_to_null($value);
         };
     }
 
-    protected function getItemId(array $record) {
-        $id = sprintf('CZE:OGL.%s_%s', $record['Rada_S'], (int)$record['PorC_S']);
+    protected function getItemId(array $record)
+    {
+        $id = sprintf('CZE:OGL.%s_%s', $record['Rada_S'], $record['PorC_S']);
         if ($record['Lomeni_S'] != '_') {
             $id = sprintf('%s-%s', $id, $record['Lomeni_S']);
         }
@@ -54,8 +55,13 @@ class OglImporter extends AbstractImporter
         return $id;
     }
 
-    protected function getItemImageFilenameFormat(array $record) {
-        $filename = sprintf('%s%s', $record['Rada_S'], str_pad($record['PorC_S'], 5, '0', STR_PAD_LEFT));
+    protected function getItemImageFilenameFormat(array $record)
+    {
+        $filename = sprintf(
+            '%s%s',
+            $record['Rada_S'],
+            str_pad($record['PorC_S'], 5, '0', STR_PAD_LEFT)
+        );
         if ($record['Lomeni_S'] != '_') {
             $filename = sprintf('%s%s', $filename, $record['Lomeni_S']);
         }
@@ -63,28 +69,51 @@ class OglImporter extends AbstractImporter
         return $filename . '{_*,}';
     }
 
-    protected function hydrateWorkType(array $record, $locale) {
+    protected function hydrateIdentifier(array $record)
+    {
+        $identifier = sprintf('%s %s', $record['Rada_S'], $record['PorC_S']);
+        if ($record['Lomeni_S'] != '_') {
+            $identifier = sprintf('%s/%s', $identifier, $record['Lomeni_S']);
+        }
+
+        return $identifier;
+    }
+
+    protected function hydrateWorkType(array $record, $locale)
+    {
         $workType = config('demus.work_type')[$record['Skupina']];
         if ($locale !== 'cs') {
-            $workType = $this->translateAttribute("work_type", $workType, $locale);
+            $workType = $this->translateAttribute('work_type', $workType, $locale);
         }
 
         return $workType;
     }
 
-    protected function hydrateTopic(array $record, $locale) {
+    protected function hydrateTopic(array $record, $locale)
+    {
         $topic = Str::lower($record['Námět']);
         if ($locale !== 'cs') {
-            $topic = $this->translateAttribute("topic", $topic, $locale);
+            $topic = $this->translateAttribute('topic', $topic, $locale);
         }
 
         return $topic;
     }
 
-    protected function hydrateTechnique(array $record, $locale) {
+    protected function hydrateStylePeriod(array $record, $locale)
+    {
+        $stylePeriod = $record['Podskup'];
+        if ($locale !== 'cs') {
+            $stylePeriod = $this->translateAttribute('style_period', $stylePeriod, $locale);
+        }
+
+        return $stylePeriod;
+    }
+
+    protected function hydrateTechnique(array $record, $locale)
+    {
         $technika = $record['Technika'];
         if ($locale !== 'cs') {
-            $technika = $this->translateAttribute("technique", $record['Technika'], $locale);
+            $technika = $this->translateAttribute('technique', $technika, $locale);
         }
 
         if (!$record['TechSpec']) {
@@ -93,22 +122,48 @@ class OglImporter extends AbstractImporter
 
         $techSpec = $record['TechSpec'];
         if ($locale !== 'cs') {
-            $techSpec = $this->translateAttribute("technique", $techSpec, $locale);
+            $techSpec = $this->translateAttribute('technique', $techSpec, $locale);
         }
 
-        if (!$technika || !$techSpec) {
-            return null;
-        }
-
-        return "$technika, $techSpec";
+        return collect([$technika, $techSpec])
+            ->filter()
+            ->join(', ') ?:
+            null;
     }
 
-    protected function hydrateMeasurement(array $record, $locale) {
+    protected function hydrateMedium(array $record, $locale)
+    {
+        $material = $record['Materiál'];
+        if ($locale !== 'cs') {
+            $material = $this->translateAttribute('medium', $record['Materiál'], $locale);
+        }
+
+        if (!$record['MatSpec']) {
+            return $material;
+        }
+
+        $matSpec = $record['MatSpec'];
+        if ($locale !== 'cs') {
+            $matSpec = $this->translateAttribute('medium', $record['MatSpec'], $locale);
+        }
+
+        return collect([$material, $matSpec])
+            ->filter()
+            ->join(', ') ?:
+            null;
+    }
+
+    protected function hydrateMeasurement(array $record, $locale)
+    {
         if (empty($record['SlužF']) || $record['SlužF'] === '=') {
             return null;
         }
 
-        $replacements = $this->translator->get('importer.demus.measurement_replacements', [], $locale);
+        $replacements = $this->translator->get(
+            'importer.demus.measurement_replacements',
+            [],
+            $locale
+        );
         return strtr($record['SlužF'], $replacements);
     }