+++ resources/lang/en/dielo.php
@@ -14,6 +14,7 @@ return array(
     'item_download'      => 'download',
 
     'item_attr_dating'         => 'date',
+    'item_attr_style_period'   => 'style period',
     'item_attr_measurements'   => 'measurements',
     'item_attr_work_type'      => 'work type',
     'item_attr_object_type'    => 'object type',
@@ -27,6 +28,7 @@ return array(
     'item_attr_integrity_work' => 'in a series with',
     'item_attr_inscription'    => 'inscription',
     'item_attr_gallery'        => 'institution',
+    'item_attr_current_location' => 'current location',
     'item_attr_contributor'    => 'curator',
     'item_attr_identifier'     => 'inventory number',
     'item_attr_licence'        => 'licence',